requirements.txt
djangorestframework==3.16.1
python-decouple==3.8
dj-database-url==3.0.1
psycopg2-binary==2.9.11
Django==6.0.2
django-filter==25.2
djrestframework-simplejwt==1.22.0
